(refs #440) Fix getRecordListForStats selector to combine records with patients and move selectors around it
