+++ src/selectors/stats.js
@@ -2,6 +2,9 @@
 
 import { createSelector } from 'reselect';
 import moment from 'moment';
+import {
+  getPatientList,
+} from './patient-list';
 
 export const getRecordList = (state: Object): Array<RecordObject> => state.recordList;
 
@@ -12,7 +15,7 @@ export const getStatsDate = (
   endDate: ?moment.Moment
 } => state.stats.date;
 
-export const getRecordListForStats = createSelector(
+export const getRecordListFilteredByDate = createSelector(
   [getRecordList, getStatsDate],
   (recordList, date) => {
     if (date.startDate == null && date.endDate == null) return recordList;
@@ -29,4 +32,20 @@ export const getRecordListForStats = createSelector(
 
     return filteredRecordList;
   }
+);
+
+export const getRecordListForStats = createSelector(
+  [getRecordListFilteredByDate, getPatientList],
+  (recordList, patientList) => recordList.map(record => {
+    const match = record._id.match(/record_(\S+)_(\S+)/);
+    const patientId = match && `patient_${match[1]}`;
+    const patient = patientId && patientList.find(p => p._id === patientId);
+
+    if (!patient) return record;
+
+    return {
+      ...record,
+      _patient: patient,
+    };
+  })
 );
